pyproject.toml
[tool.poetry]
name = "chinese-address-parser"
version = "0.1.0"
description = ""
authors = ["Goldyban Nastya <nasty-goldyba@ya.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2"
djangorestframework = "^3.16.0"
psycopg2 = "^2.9.10"
python-dotenv = "^1.1.0"
django-cors-headers = "^4.7.0"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.2.0"
black = "^25.1.0"
isort = "^6.0.1"


[tool.poetry.group.dev.dependencies]
coverage = "^7.8.0"

[tool.black]
line-length = 119
exclude = '.git'

[tool.isort]
line_length = 119

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
